Accès membre/admin

+++ commun/nav.php
@@ -2,9 +2,17 @@
 
 <ul class="nav col-12 col-lg-auto me-lg-auto mb-2 justify-content-center mb-md-0">
   <li><a href="index.php" class="nav-link px-2 text-white">index</a></li>
-  <li><a href="add_meal.php" class="nav-link px-2 text-white">Ajouter un repas</a></li>
-  <li><a href="add_ingredients.php" class="nav-link px-2 text-white">Ajouter un ingredients</a></li>
-  <li><a href="deletetable.php" class="nav-link px-2 text-white">Delete table</a></li>
+  <?php
+    if (isset($_SESSION['auth'])) { ?>
+      <li><a href="add_meal.php" class="nav-link px-2 text-white">Ajouter un repas</a></li>
+      <li><a href="add_ingredients.php" class="nav-link px-2 text-white">Ajouter un ingredients</a></li>
+  
+      <?php
+        if ($_SESSION['role']==1) { ?>
+          <li><a href="deletetable.php" class="nav-link px-2 text-white">Delete table</a></li>   
+        <?php
+        }
+    } ?>
 </ul>
 </nav>
 <form class="col-12 col-lg-auto mb-3 mb-lg-0 me-lg-3" role="search">
